Change *int32 to *int64

+++ pkg/apis/kubeflow/v1alpha1/types.go
@@ -57,7 +57,7 @@ type MPIJobSpec struct {
 	// the job may be active before the system tries to terminate it.
 	// Note that this takes precedence over `BackoffLimit` field.
 	// +optional
-	ActiveDeadlineSeconds *int32 `json:"activeDeadlineSeconds,omitempty"`
+	ActiveDeadlineSeconds *int64 `json:"activeDeadlineSeconds,omitempty"`
 
 	// Specifies the desired number of replicas the MPIJob should run on.
 	// The `PodSpec` should specify the number of GPUs.
